+++ pom.xml
@@ -10,7 +10,6 @@
     </parent>
 
     <artifactId>vaadin-iron-list-flow</artifactId>
-    <version>1.0-SNAPSHOT</version>
 
     <properties>
         <flow.version>1.0-SNAPSHOT</flow.version>
